refactor(common): move ProtoSerializer (PR change request)
